+++ src/save-structure/const-data/index.ts
@@ -1,5 +1,6 @@
 export * from "./accessories";
 export * from "./geysers";
 export * from "./skills";
+export * from "./space-destinations";
 
 export * from "./template-enumerations";