+++ src/utils.ts
@@ -41,3 +41,7 @@ export function validateDotNetIdentifierName(
 export function typed<T extends string>(s: T): T {
   return s;
 }
+
+export function typedKeys<T>(x: T): (keyof T)[] {
+  return Object.keys(x) as any;
+}